move trainer config
