+++ examples/backpack_example.py
@@ -15,14 +15,13 @@ import wandb
 from haliax import Axis
 from haliax.partitioning import ResourceAxis, named_jit, round_axis_for_partitioning
 from levanter import callbacks
-from levanter.config import OptimizerConfig, TrainerConfig
 from levanter.data.loader import ReplicatedBatchLoader, ShardedBatchLoader
 from levanter.data.text import LMDatasetConfig, TokenSeqDataset
 from levanter.grad_accum import accumulate_gradients_sharded
 from levanter.logging import capture_time, log_time_to_wandb
 from levanter.models.backpack import BackpackConfig, BackpackLMHeadModel
 from levanter.models.loss import next_token_loss
-from levanter.trainer import StepInfo, TrainerHooks
+from levanter.trainer import OptimizerConfig, StepInfo, TrainerConfig, TrainerHooks
 from levanter.utils.jax_utils import global_key_array, parameter_count
 from levanter.utils.py_utils import non_caching_cycle
