+++ examples/mpt_continued_training.py
@@ -15,7 +15,6 @@ import wandb
 from haliax import Axis
 from haliax.partitioning import ResourceAxis, named_jit
 from levanter import callbacks
-from levanter.config import OptimizerConfig, TrainerConfig
 from levanter.data.sharded import GlobalBatchDataset, LocalBatchDataset
 from levanter.data.text import LMDatasetConfig, TokenSeqDataset
 from levanter.grad_accum import accumulate_gradients_sharded
@@ -23,7 +22,7 @@ from levanter.logging import capture_time, log_time_to_wandb
 from levanter.models.gpt2 import Gpt2LMHeadModel
 from levanter.models.loss import next_token_loss
 from levanter.models.mpt import MptLmHeadModel
-from levanter.trainer import StepInfo, TrainerHooks
+from levanter.trainer import OptimizerConfig, StepInfo, TrainerConfig, TrainerHooks
 from levanter.utils.jax_utils import global_key_array, parameter_count
 from levanter.utils.py_utils import non_caching_cycle
